external/webkit/LayoutTests/platform/android/webgl-conformance/dumpAsText/conformance/glsl/misc/shader-with-function-recursion.frag-expected.txt

CONSOLE MESSAGE: line 24: *** Error compiling shader '[object WebGLShader]':ERROR: Function recursion detected

shader with recursive function calls should fail

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

test: shader with recursive function calls should fail
default vertex shader
test fragment shaderPASS shader with recursive function calls should fail
PASS successfullyParsed is true

TEST COMPLETE
